Move braille into separate file
