+++ Cargo.toml
@@ -41,14 +41,8 @@ rayon = { version = "1", optional = true }
 default = ["full"]
 full = ["rayon", "player"]
 
-player = [
-    "dep:clap",
-    "dep:crossterm",
-    "dep:indicatif",
-    "dep:glob",
-    "dep:imageproc",
-]
-
+braille = ["dep:imageproc"]
+player = ["braille", "dep:clap", "dep:crossterm", "dep:indicatif", "dep:glob"]
 rayon = ["dep:rayon", "imageproc/rayon", "image/rayon", "indicatif/rayon"]
 multithreading = ["rayon"]
 parallelism = ["rayon"]
